dependencies/BC_Policy/requirements.txt
glog==0.3.1
numpy==2.1.2
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.6.77
opencv-python==4.12.0.88
pandas==2.3.3
pillow==11.3.0
platformdirs==4.5.0
prompt_toolkit==3.0.51
protobuf==6.32.1
ptyprocess==0.7.0
threadpoolctl==3.6.0
torch==2.9.0+cu126
torchvision==0.24.0+cu126
tqdm==4.67.1
transforms3d==0.4.2
wandb==0.22.3
xacro==2.0.13
